🚧 0SFMS7 task: isolate evaluator runtime evidence test
